refactor: extract tests from generate_sbom.rs into separate module
Split generate_sbom.rs (1397 lines) into a directory module structure:
- generate_sbom/mod.rs: Implementation code (~534 lines)
- generate_sbom/tests.rs: Test code (~860 lines)

This improves code navigation and maintainability by separating
implementation from test code, following Rust's standard module
organization pattern.

Closes #125
